+++ src/ts/functions/objects/makeNumberAsync.ts
@@ -1,5 +1,5 @@
 /**
- * @since ___PKG_VERSION___
+ * @since 2.0.0-beta.2.draft
  * 
  * @packageDocumentation
  */
@@ -14,7 +14,7 @@ import { makeNumber } from './makeNumber.js';
  * Makes any input a number if it can be made into a valid number (e.g., by
  * parsing a string).
  * 
- * @since ___PKG_VERSION___
+ * @since 2.0.0-beta.2.draft
  */
 export async function makeNumberAsync( input: unknown ): Promise<null | number> {
     return makeNumber( input );