+++ src/ts/functions/objects/objectMap.ts
@@ -1,5 +1,5 @@
 /**
- * @since ___PKG_VERSION___
+ * @since 2.0.0-beta.2.draft
  * 
  * @packageDocumentation
  */
@@ -14,7 +14,7 @@
  * @param obj     The object to map.
  * @param mapper  The callback function used to define new values.
  * 
- * @since ___PKG_VERSION___
+ * @since 2.0.0-beta.2.draft
  */
 export function objectMap<
     T_Obj extends object,